backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
openai
google-genai
qdrant-client
python-dotenv
